refactor(schema): preserve event domain ownership
