+++ packages/core/src/public-event-manifest.ts
@@ -1,7 +1,6 @@
 export * as PublicEventManifest from "./public-event-manifest"
 
 import { Event } from "@opencode-ai/schema/event"
-import { Todo } from "@opencode-ai/schema/todo"
 import { Catalog } from "./catalog"
 import { FileSystem } from "./filesystem"
 import { Watcher } from "./filesystem/watcher"
@@ -14,6 +13,7 @@ import { Pty } from "./pty"
 import { QuestionV2 } from "./question"
 import { Reference } from "./reference"
 import { EventManifest } from "./event-manifest"
+import { SessionTodo } from "./session/todo"
 
 export const FoundationDefinitions = Event.inventory(
   ModelsDev.Event.Refreshed,
@@ -38,10 +38,11 @@ export const FeatureDefinitions = Event.inventory(
   QuestionV2.Event.Asked,
   QuestionV2.Event.Replied,
   QuestionV2.Event.Rejected,
-  Todo.Event.Updated,
 )
 
-export const Definitions = Event.inventory(...FoundationDefinitions, ...FeatureDefinitions)
+export const TodoDefinitions = Event.inventory(SessionTodo.Event.Updated)
+
+export const Definitions = Event.inventory(...FoundationDefinitions, ...FeatureDefinitions, ...TodoDefinitions)
 
 export const Latest = Event.latest(Definitions)
 export const Durable = Event.durable(Definitions)
