+++ packages/core/src/session/todo.ts
@@ -1,16 +1,30 @@
 export * as SessionTodo from "./todo"
 
 import { asc, eq } from "drizzle-orm"
-import { Context, Effect, Layer } from "effect"
-import { Todo } from "@opencode-ai/schema/todo"
+import { Context, Effect, Layer, Schema } from "effect"
 import { Database } from "../database/database"
 import { EventV2 } from "../event"
 import { SessionSchema } from "./schema"
 import { TodoTable } from "./sql"
 
-export const Info = Todo.Info
-export type Info = Todo.Info
-export const Event = Todo.Event
+export const Info = Schema.Struct({
+  content: Schema.String.annotate({ description: "Brief description of the task" }),
+  status: Schema.String.annotate({
+    description: "Current status of the task: pending, in_progress, completed, cancelled",
+  }),
+  priority: Schema.String.annotate({ description: "Priority level of the task: high, medium, low" }),
+}).annotate({ identifier: "SessionTodo.Info" })
+export type Info = typeof Info.Type
+
+export const Event = {
+  Updated: EventV2.define({
+    type: "todo.updated",
+    schema: {
+      sessionID: SessionSchema.ID,
+      todos: Schema.Array(Info),
+    },
+  }),
+}
 
 export interface Interface {
   readonly update: (input: {
