+++ packages/opencode/test/event-manifest.test.ts
@@ -1,20 +1,20 @@
 import { describe, expect, test } from "bun:test"
-import { SessionEvent } from "@opencode-ai/schema/session-event"
+import { SessionEvent } from "@opencode-ai/core/session/event"
+import { Todo } from "@/session/todo"
 import { EventManifest } from "@/event-manifest"
 
 describe("public event manifest", () => {
   test("contains every latest public wire type once", () => {
-    expect(EventManifest.Latest.size).toBe(86)
+    expect(EventManifest.Latest.size).toBe(85)
     expect(EventManifest.Latest.get("session.next.step.ended")).toBe(SessionEvent.Step.Ended)
-    expect(EventManifest.Latest.has("ide.installed")).toBe(true)
+    expect(EventManifest.Latest.get("todo.updated")).toBe(Todo.Event.Updated)
+    expect(EventManifest.Latest.has("ide.installed")).toBe(false)
     expect(EventManifest.Latest.has("server.connected")).toBe(true)
     expect(EventManifest.Latest.has("global.disposed")).toBe(true)
   })
 
-  test("keeps historical durable versions out of the latest manifest", () => {
-    expect(EventManifest.Latest.values().toArray()).not.toContain(SessionEvent.Step.EndedV1)
-    expect(EventManifest.Latest.values().toArray()).not.toContain(SessionEvent.Step.FailedV1)
-    expect(EventManifest.Durable.get("session.next.step.ended.1")).toBe(SessionEvent.Step.EndedV1)
+  test("contains only the current step settlement versions", () => {
+    expect(EventManifest.Durable.has("session.next.step.ended.1")).toBe(false)
     expect(EventManifest.Durable.get("session.next.step.ended.2")).toBe(SessionEvent.Step.Ended)
   })
 })
