+++ packages/schema/src/index.ts
@@ -9,16 +9,13 @@ export { LLM } from "./llm"
 export { Location } from "./location"
 export { Model } from "./model"
 export { Permission } from "./permission"
-export { PermissionV1 } from "./permission-v1"
 export { Project } from "./project"
 export { Provider } from "./provider"
 export { Reference } from "./reference"
 export { Session } from "./session"
-export { SessionV1 } from "./session-v1"
 export { SessionInput } from "./session-input"
 export { SessionMessage } from "./session-message"
 export { Skill } from "./skill"
-export { Todo } from "./todo"
 export { Workspace } from "./workspace"
 export { Prompt, Source, FileAttachment, AgentAttachment } from "./prompt"
 export * from "./schema"
