+++ packages/sdk/js/src/v2/gen/types.gen.ts
@@ -88,7 +88,6 @@ export type Event =
   | EventWorkspaceStatus
   | EventWorktreeReady
   | EventWorktreeFailed
-  | EventIdeInstalled
   | EventServerConnected
   | EventGlobalDisposed
   | EventServerInstanceDisposed
@@ -1600,13 +1599,6 @@ export type GlobalEvent = {
       }
     | {
         id: string
-        type: "ide.installed"
-        properties: {
-          ide: string
-        }
-      }
-    | {
-        id: string
         type: "server.connected"
         properties: {
           [key: string]: unknown
@@ -2841,7 +2833,6 @@ export type V2Event =
   | V2EventWorkspaceStatus
   | V2EventWorktreeReady
   | V2EventWorktreeFailed
-  | V2EventIdeInstalled
   | V2EventServerConnected
   | V2EventGlobalDisposed
 
@@ -5911,23 +5902,6 @@ export type V2EventWorktreeFailed = {
   }
 }
 
-export type V2EventIdeInstalled = {
-  id: string
-  metadata?: {
-    [key: string]: unknown
-  }
-  durable?: {
-    aggregateID: string
-    seq: number
-    version: number
-  }
-  location?: LocationRef
-  type: "ide.installed"
-  data: {
-    ide: string
-  }
-}
-
 export type V2EventServerConnected = {
   id: string
   metadata?: {
@@ -6899,14 +6873,6 @@ export type EventWorktreeFailed = {
   type: "worktree.failed"
   properties: {
     message: string
-  }
-}
-
-export type EventIdeInstalled = {
-  id: string
-  type: "ide.installed"
-  properties: {
-    ide: string
   }
 }
