+++ packages/server/package.json
@@ -13,7 +13,6 @@
   },
   "dependencies": {
     "@opencode-ai/core": "workspace:*",
-    "@opencode-ai/schema": "workspace:*",
     "drizzle-orm": "catalog:",
     "effect": "catalog:"
   },
