+++ packages/server/src/api.ts
@@ -9,7 +9,7 @@ import { FileSystemGroup } from "./groups/fs"
 import { CommandGroup } from "./groups/command"
 import { SkillGroup } from "./groups/skill"
 import { EventGroup, makeEventGroup } from "./groups/event"
-import type { Definition } from "@opencode-ai/schema/event"
+import type { Definition } from "@opencode-ai/core/event"
 import { AgentGroup } from "./groups/agent"
 import { HealthGroup } from "./groups/health"
 import { PtyGroup } from "./groups/pty"
